+++ docs/project__operations_8cc.js
@@ -1,0 +1,6 @@
+var project__operations_8cc =
+[
+    [ "close_projectc", "project__operations_8cc.html#ae082f93bce8bf519804af78cc9f0e4f7", null ],
+    [ "config_open_projectc", "project__operations_8cc.html#a670c23ed754f6cfa811b357c09935538", null ],
+    [ "open_projectc", "project__operations_8cc.html#a1a505b76764b166689fe978433d60bcb", null ]
+];