+++ docs/search/all_8.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['node_2djs_20bindings_52',['Node-js bindings',['../index.html',1,'']]]
+];