+++ docs/search/files_4.js
@@ -1,0 +1,4 @@
+var searchData=
+[
+  ['object_5foperations_2ecc_98',['object_operations.cc',['../object__operations_8cc.html',1,'']]]
+];