+++ docs/upload__operations_8cc.js
@@ -1,0 +1,9 @@
+var upload__operations_8cc =
+[
+    [ "upload_abortc", "upload__operations_8cc.html#acec0b997c68975ed65af14cc200bad09", null ],
+    [ "upload_commitc", "upload__operations_8cc.html#ab5167c4273b9e19024e9d401de4f3605", null ],
+    [ "upload_infoc", "upload__operations_8cc.html#a46ee7742d3930e8dd98064b8d1a504b9", null ],
+    [ "upload_objectc", "upload__operations_8cc.html#af5f30faecbe174c8a9ebf8957c6d8f8c", null ],
+    [ "upload_set_custom_metadatac", "upload__operations_8cc.html#aa93e4e71b750bc9f2ece0e705be27628", null ],
+    [ "upload_writec", "upload__operations_8cc.html#aa2a3753dc8e5ac1f26a27b0f6b55738c", null ]
+];