+++ functions_win/bucket_operations_win.h
@@ -1,0 +1,11 @@
+#include "definitions_win.h"
+#include "../libUplink_win.h"
+
+napi_value stat_bucketc(napi_env env, napi_callback_info info);
+//
+napi_value create_bucketc(napi_env env, napi_callback_info info);
+//
+napi_value ensure_bucketc(napi_env env, napi_callback_info info);
+//
+napi_value delete_bucketc(napi_env env, napi_callback_info info);
+//