+++ functions_win/object_operations_win.h
@@ -1,0 +1,6 @@
+#include "definitions_win.h"
+#include "../libUplink_win.h"
+
+napi_value stat_objectc(napi_env env, napi_callback_info info);
+//
+napi_value delete_objectc(napi_env env, napi_callback_info info);