+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Storj-Nodejs Changelog
 
+## [1.0.5] - 29-06-2020
+* Bugs fixed
+* Removed docs folder
+
+## [1.0.4] - 25-06-2020
+* Asyncronous function
+* Changed HelloStorj.js
+* Splitted c files
+
 ## [1.0.3] - 09-06-2020
 ### Changelog:
 * Updated ReadMe